DTSPB-3402 Removed unnecesssary api-utils
