+++ app/services/Service.js
@@ -6,7 +6,6 @@ const logger = require('app/components/logger');
 const config = require('config');
 const formatUrl = require('app/utils/FormatUrl');
 const AsyncFetch = require('app/utils/AsyncFetch');
-const asyncFetch = new AsyncFetch();
 
 class Service {
     constructor(endpoint, sessionId) {
@@ -42,21 +41,21 @@ class Service {
     }
 
     fetchJson(url, fetchOptions) {
-        return asyncFetch
+        return AsyncFetch
             .fetch(url, fetchOptions, res => res.json())
             .then(json => json)
             .catch(err => err);
     }
 
     fetchText(url, fetchOptions) {
-        return asyncFetch
+        return AsyncFetch
             .fetch(url, fetchOptions, res => res.text())
             .then(text => text)
             .catch(err => err);
     }
 
     fetchBuffer(url, fetchOptions) {
-        return asyncFetch
+        return AsyncFetch
             .fetch(url, fetchOptions, res => res.buffer())
             .then(buffer => buffer)
             .catch(err => {
